requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────
numpy>=2.1,<2.3
pandas>=2.2,<2.4
scipy>=1.14,<1.16
scikit-learn>=1.6,<1.8

# ── Deep Learning ────────────────────────────────────────────────────────
tensorflow>=2.21,<2.22

# ── MLflow ───────────────────────────────────────────────────────────────
mlflow>=2.13

# ── API ──────────────────────────────────────────────────────────────────
fastapi>=0.111,<0.116
uvicorn[standard]>=0.29,<0.35
pydantic>=2.7,<2.12

# ── Utilidades ───────────────────────────────────────────────────────────
matplotlib>=3.9,<3.11
seaborn>=0.13,<0.14
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
httpx>=0.27,<0.29